4 - Chart v2

+++ restapi/src/main/java/com/felipeshiba/bitcoin/restapi/chart/ValueMarketPriceChartResponse.kt
@@ -1,6 +1,6 @@
 package com.felipeshiba.bitcoin.restapi.chart
 
 data class ValueMarketPriceChartResponse(
-    val x: Int,
+    val x: Long,
     val y: Double
 )
